+++ src/ca/uwindsor/css/ctf/PlayerJoinListener.java
@@ -1,4 +1,4 @@
-package com.pickuperic;
+package ca.uwindsor.css.ctf;
 
 import org.bukkit.entity.Player;
 import org.bukkit.event.EventHandler;
