+++ src/ca/uwindsor/css/ctf/SafezoneListener.java
@@ -1,4 +1,4 @@
-package com.pickuperic;
+package ca.uwindsor.css.ctf;
 
 import java.util.Hashtable;
