+++ src/ca/uwindsor/css/ctf/Safezones.java
@@ -1,4 +1,4 @@
-package com.pickuperic;
+package ca.uwindsor.css.ctf;
 
 import java.util.ArrayList;
