+++ src/ca/uwindsor/css/ctf/Teams.java
@@ -1,4 +1,4 @@
-package com.pickuperic;
+package ca.uwindsor.css.ctf;
 
 import java.util.ArrayList;
 import java.util.Collection;
